usb: gadget: Send zero-length packets on bulk IN
Set the flag that enables the usb device controller to
send zero-length packet if the payload size is a multiple
of the max packet size. For such payload size if zero length
packet is not sent, IP packet received by host driver after
transfer completion will be a corrupted IP packet which will
be dropped by network layer. This results in downlink data
stall.

Change-Id: Ica99e0be7f18f27f7a77c31526f51d88df4682ec
Signed-off-by: Hemant Kumar <hemantk@codeaurora.org>
Signed-off-by: Seongmin Park <r_data@naver.com>