+ "Optimise apply_slack() for size and speed." - http://lkml.indiana.edu/hypermail/linux/kernel/1201.3/03172.html
To apply proper slack, the original algorithm used to prepare a mask and
then apply the mask to obtain the appropriately rounded-off absolute
time the timer expires.

This patch modifies the masking logic to a bit-shift logic, therby
reducing the complexity and number of operations. Thus obtaining a minor
speed-up.